+++ metadata.rb
@@ -10,8 +10,6 @@ issues_url        'https://github.com/sous-chefs/postgresql/issues'
 chef_version      '>= 15.3'
 
 depends 'yum', '>= 7.2.0'
-depends 'yum-epel'
-depends 'apt'
 
 supports 'amazon'
 supports 'centos'