+++ metadata.rb
@@ -4,23 +4,18 @@ maintainer        'Sous Chefs'
 maintainer_email  'help@sous-chefs.org'
 license           'Apache-2.0'
 description       'Installs and configures postgresql for clients or servers'
-version           '9.0.3'
+version           '9.1.0'
 source_url        'https://github.com/sous-chefs/postgresql'
 issues_url        'https://github.com/sous-chefs/postgresql/issues'
 chef_version      '>= 16.0'
 
-depends 'yum-epel'
-depends 'apt'
+depends 'yum', '>= 7.2.0'
 
-%w(
-  amazon
-  centos
-  debian
-  fedora
-  oracle
-  redhat
-  scientefic
-  ubuntu
-).each do |os|
-  supports os
-end
+supports 'amazon'
+supports 'centos'
+supports 'debian'
+supports 'fedora'
+supports 'oracle'
+supports 'redhat'
+supports 'scientific'
+supports 'ubuntu'