+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 This file is used to list changes made in the last 3 major versions of the postgresql cookbook.
 
+## Unreleased
+
 ## 9.0.3 - *2021-08-30*
 
 - Standardise files with files in sous-chefs/repo-management