+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 This file is used to list changes made in the last 3 major versions of the postgresql cookbook.
 
+## Unreleased
+
 ## 9.1.0 - *2021-10-07*
 
 - Use `dnf_module` resource from yum cookbook instead of manually shelling out