+++ CHANGELOG.md
@@ -7,6 +7,13 @@ This file is used to list changes made in the last 3 major versions of the postg
 - Move connection options to a resource partial
 - Require Chef 16+ for resource partials
 
+## 9.1.0 - *2021-10-07*
+
+- Use `dnf_module` resource from yum cookbook instead of manually shelling out
+  - requires new dependency on `yum` >= 7.2.0
+- Remove unneeded `apt` & `yum-epel` dependencies
+- Move resource documentation out of README
+
 ## 9.0.3 - *2021-08-30*
 
 - Standardise files with files in sous-chefs/repo-management